fix: move controls to one file
