+++ src/components/media/controls/controls.css
@@ -1,3 +1,32 @@
+.rustic-progress {
+  display: flex;
+  gap: 24px;
+  align-items: center;
+  position: relative;
+}
+
+.rustic-progress-slider {
+  height: 8px;
+}
+
+/* circle that you drag on slider to change the time */
+.rustic-progress-slider .MuiSlider-thumb {
+  z-index: 2;
+  width: 15px;
+  height: 15px;
+}
+/* bar to the left of the thumb (circle), tracking the time */
+.rustic-progress-slider .MuiSlider-track {
+  z-index: 1;
+}
+
+.rustic-progress-buffered {
+  height: 8px;
+  position: absolute;
+  border-radius: 12px;
+  pointer-events: none;
+}
+
 .rustic-volume-settings {
   display: flex;
   align-items: center;
@@ -22,6 +51,14 @@
   display: unset;
 }
 
+.rustic-transcript-toggle span {
+  text-transform: none;
+}
+
+.rustic-playback-rate-button {
+  min-width: 0;
+}
+
 @media (min-width: 900px) {
   /* slider appears on hover or focus */
   .rustic-volume-settings:hover .rustic-volume-slider,
@@ -32,6 +69,14 @@
 }
 
 @media (max-width: 900px) {
+  .rustic-progress {
+    padding-inline: 0;
+  }
+
+  .rustic-progress-buffered {
+    left: 0;
+  }
+
   .rustic-volume-settings {
     gap: 8px;
   }
@@ -43,5 +88,9 @@
 
   .rustic-volume-settings .MuiSlider-thumb {
     display: unset;
+  }
+
+  .rustic-pause-play-icon {
+    font-size: 52px;
   }
 }
