--- src/components/media/pausePlayToggle/pausePlayToggle.css
@@ -1,5 +1,0 @@
-@media (max-width: 900px) {
-  .rustic-pause-play-icon {
-    font-size: 52px;
-  }
-}
