--- src/components/media/playbackRateButton/playbackRateButton.css
@@ -1,3 +1,0 @@
-.rustic-playback-rate-button {
-  min-width: 0;
-}
