--- src/components/media/progressSlider/progressSlider.css
@@ -1,38 +1,0 @@
-.rustic-progress {
-  display: flex;
-  gap: 24px;
-  align-items: center;
-  position: relative;
-}
-
-.rustic-progress-slider {
-  height: 8px;
-}
-
-/* circle that you drag on slider to change the time */
-.rustic-progress-slider .MuiSlider-thumb {
-  z-index: 2;
-  width: 15px;
-  height: 15px;
-}
-/* bar to the left of the thumb (circle), tracking the time */
-.rustic-progress-slider .MuiSlider-track {
-  z-index: 1;
-}
-
-.rustic-progress-buffered {
-  height: 8px;
-  position: absolute;
-  border-radius: 12px;
-  pointer-events: none;
-}
-
-@media (max-width: 900px) {
-  .rustic-progress {
-    padding-inline: 0;
-  }
-
-  .rustic-progress-buffered {
-    left: 0;
-  }
-}
