--- src/components/media/transcriptToggle/transcriptToggle.css
@@ -1,3 +1,0 @@
-.rustic-transcript-toggle span {
-  text-transform: none;
-}
